tech(api): migrate route / controller
