+++ api/lib/application/sco-organization-learners/index.js
@@ -22,37 +22,6 @@ const inaPattern = new RegExp('^[0-9]{10}[a-zA-Z]{1}$');
 const register = async function (server) {
   server.route([
     {
-      method: 'POST',
-      path: '/api/sco-organization-learners/association',
-      config: {
-        handler: scoOrganizationLearnerController.reconcileScoOrganizationLearnerManually,
-        validate: {
-          options: {
-            allowUnknown: false,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                'first-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                'last-name': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-                birthdate: Joi.date().format('YYYY-MM-DD').required(),
-                'campaign-code': Joi.string().empty(Joi.string().regex(/^\s*$/)).required(),
-              },
-              type: 'sco-organization-learners',
-            },
-          }),
-          failAction: (request, h) => {
-            return sendJsonApiError(new UnprocessableEntityError('Un des champs saisis n’est pas valide.'), h);
-          },
-        },
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Elle associe des données saisies par l’utilisateur à l’inscription de l’élève dans cette organisation',
-        ],
-        tags: ['api', 'sco-organization-learners'],
-      },
-    },
-    {
       method: 'PUT',
       path: '/api/sco-organization-learners/possibilities',
       config: {
