+++ api/tests/prescription/learner-management/acceptance/application/sco-organization-management-route_test.js
@@ -1,0 +1,112 @@
+import {
+  createServer,
+  databaseBuilder,
+  expect,
+  generateValidRequestAuthorizationHeader,
+  knex,
+} from '../../../../test-helper.js';
+
+describe('Acceptance | Route | sco-organization-management-route', function () {
+  let server;
+
+  beforeEach(async function () {
+    server = await createServer();
+  });
+
+  describe('POST /api/sco-organization-learners/association', function () {
+    let organization;
+    let campaign;
+    let options;
+    let organizationLearner;
+    let user;
+
+    beforeEach(async function () {
+      // given
+      options = {
+        method: 'POST',
+        url: '/api/sco-organization-learners/association',
+        headers: {},
+        payload: {},
+      };
+
+      user = databaseBuilder.factory.buildUser();
+      organization = databaseBuilder.factory.buildOrganization({ type: 'SCO' });
+      organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
+        firstName: 'france',
+        lastName: 'gall',
+        birthdate: '2001-01-01',
+        organizationId: organization.id,
+        userId: null,
+        nationalStudentId: 'francegall123',
+      });
+      campaign = databaseBuilder.factory.buildCampaign({ organizationId: organization.id });
+
+      await databaseBuilder.commit();
+    });
+
+    context('associate user with firstName, lastName and birthdate', function () {
+      it('should return an 200 status after having successfully associated user to organizationLearner', async function () {
+        // given
+        options.headers.authorization = generateValidRequestAuthorizationHeader(user.id);
+        options.payload.data = {
+          attributes: {
+            'campaign-code': campaign.code,
+            'first-name': organizationLearner.firstName,
+            'last-name': organizationLearner.lastName,
+            birthdate: organizationLearner.birthdate,
+          },
+        };
+        options.url += '?withReconciliation=true';
+
+        // when
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+      });
+
+      context('when user is not authenticated', function () {
+        it('should respond with a 401 - unauthorized access', async function () {
+          // given
+          options.headers.authorization = 'invalid.access.token';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(401);
+        });
+      });
+
+      context('When withReconciliation query param is set to false', function () {
+        it('should not reconcile user and return a 204 No Content', async function () {
+          // given
+          options.headers.authorization = generateValidRequestAuthorizationHeader(user.id);
+          options.payload.data = {
+            attributes: {
+              'campaign-code': campaign.code,
+              'first-name': organizationLearner.firstName,
+              'last-name': organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            },
+          };
+          options.url += '?withReconciliation=false';
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(204);
+          const organizationLearnerInDB = await knex('organization-learners')
+            .where({
+              firstName: organizationLearner.firstName,
+              lastName: organizationLearner.lastName,
+              birthdate: organizationLearner.birthdate,
+            })
+            .select();
+          expect(organizationLearnerInDB.userId).to.be.undefined;
+        });
+      });
+    });
+  });
+});
